Simplify BibDatabaseWriter by removing unnecessary subclass
