+++ jablib/src/jmh/java/org/jabref/benchmarks/Benchmarks.java
@@ -10,7 +10,7 @@ import org.jabref.logic.bibtex.FieldPreferences;
 import org.jabref.logic.citationkeypattern.CitationKeyPatternPreferences;
 import org.jabref.logic.exporter.BibDatabaseWriter;
 import org.jabref.logic.exporter.BibWriter;
-import org.jabref.logic.exporter.BibtexDatabaseWriter;
+import org.jabref.logic.exporter.BibDatabaseWriter;
 import org.jabref.logic.exporter.SelfContainedSaveConfiguration;
 import org.jabref.logic.formatter.bibtexfields.HtmlToLatexFormatter;
 import org.jabref.logic.importer.ParserResult;
@@ -83,7 +83,7 @@ public class Benchmarks {
         FieldPreferences fieldPreferences = new FieldPreferences(true, List.of(), List.of());
         CitationKeyPatternPreferences citationKeyPatternPreferences = mock(CitationKeyPatternPreferences.class, Answers.RETURNS_DEEP_STUBS);
 
-        BibtexDatabaseWriter databaseWriter = new BibtexDatabaseWriter(
+        BibDatabaseWriter databaseWriter = new BibDatabaseWriter(
                 bibWriter,
                 saveConfiguration,
                 fieldPreferences,
