+++ jablib/src/test/java/org/jabref/logic/importer/fileformat/BibtexImporterTest.java
@@ -34,7 +34,7 @@ import static org.mockito.Mockito.mock;
 /**
  * This class tests the BibtexImporter.
  * <p>
- * Tests for writing can be found at {@link org.jabref.logic.exporter.BibtexDatabaseWriterTest}.
+ * Tests for writing can be found at {@link org.jabref.logic.exporter.BibDatabaseWriterTest}.
  * Tests for parsing single entry BibTeX can be found at {@link BibtexParserTest}
  */
 class BibtexImporterTest {
