+++ jablib/src/test/java/org/jabref/logic/pseudonymization/PseudonymizationTest.java
@@ -12,7 +12,7 @@ import org.jabref.logic.bibtex.FieldPreferences;
 import org.jabref.logic.citationkeypattern.CitationKeyPatternPreferences;
 import org.jabref.logic.exporter.BibDatabaseWriter;
 import org.jabref.logic.exporter.BibWriter;
-import org.jabref.logic.exporter.BibtexDatabaseWriter;
+import org.jabref.logic.exporter.BibDatabaseWriter;
 import org.jabref.logic.exporter.SelfContainedSaveConfiguration;
 import org.jabref.logic.importer.ImportFormatPreferences;
 import org.jabref.logic.importer.fileformat.BibtexImporter;
@@ -38,7 +38,7 @@ class PseudonymizationTest {
 
     private BibtexImporter importer;
 
-    private BibtexDatabaseWriter databaseWriter;
+    private BibDatabaseWriter databaseWriter;
     private StringWriter stringWriter;
     private BibWriter bibWriter;
     private SelfContainedSaveConfiguration saveConfiguration;
@@ -57,7 +57,7 @@ class PseudonymizationTest {
         citationKeyPatternPreferences = mock(CitationKeyPatternPreferences.class, Answers.RETURNS_DEEP_STUBS);
         entryTypesManager = new BibEntryTypesManager();
 
-        databaseWriter = new BibtexDatabaseWriter(
+        databaseWriter = new BibDatabaseWriter(
                 bibWriter,
                 saveConfiguration,
                 fieldPreferences,
